FLASK/requirements.txt
blinker==1.8.2
click==8.1.7
contourpy==1.2.1
cv==1.0.0
cycler==0.12.1
Flask==3.0.3
fonttools==4.51.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
libigl==2.5.1
MarkupSafe==2.1.5
matplotlib==3.9.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pymesh==1.0.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
threadpoolctl==3.5.0
trimesh==4.4.0
Werkzeug==3.0.3
